react_mcp/pyproject.toml
[project]
name = "react-mcp"
version = "0.1.0"
description = ""
authors = [
    {name = "sungminna",email = "sungmin.na330@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13,<4.0"
dependencies = [
    "langchain-mcp-adapters (>=0.0.7,<0.0.8)",
    "langgraph (>=0.3.27,<0.4.0)",
    "langchain-openai (>=0.3.12,<0.4.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "fastapi[standard] (>=0.115.12,<0.116.0)",
    "passlib[bcrypt] (>=1.7.4,<2.0.0)",
    "uvicorn (>=0.34.1,<0.35.0)",
    "sqlalchemy (>=2.0.40,<3.0.0)",
    "python-jose[cryptography] (>=3.4.0,<4.0.0)",
    "python-multipart (>=0.0.20,<0.0.21)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
